requirements.txt
# This file is auto-generated from pyproject.toml
# For development, use: poetry install

streamlit>=1.31
pillow>=10.2
openai>=1.12
elevenlabs>=1.0
moviepy>=1.0
pydub>=0.25
replicate>=0.25
python-dotenv>=1.0
requests>=2.31
numpy>=1.26
gtts>=2.5
fastapi>=0.109
uvicorn[standard]>=0.27

# Development dependencies
pytest>=7.0
pytest-cov>=4.0
black>=24.0
pre-commit>=3.0
ruff>=0.7
